requirements.txt
asgiref==3.5.0
Django==3.2.12
pytz==2021.3
sqlparse==0.4.2

#Heroku
django-heroku==0.0.0

#allauth
django-allauth==0.49.0

#crispy
django-crispy-forms==1.14.0

#pytest
pytest-django
pytest-factoryboy
pytest-cov

#dotenv
python-dotenv==0.19.2



